Mehr Robustheit im Fall von Konfigurationsfehlern oder überlangen Diffs

+++ deploy.sh
@@ -70,6 +70,13 @@ cp -a "$root/html/galleries" luga-dummy/
 # wget holt natürlich nicht Ressourcen, die nur von JavaScript aus referenziert
 # werden. Daher ist eine manuelle Kopie der JavaScript-Gallerien nötig.
 
+if [ ! -e luga-dummy/index.html ]; then
+    echo "Didn't manage to mirror 'index.html'; something went wrong. Aborting." >&2
+    echo "$ curl -v http://luga-dummy/" >&2
+    curl -v http://luga-dummy/ >&2 || true
+    exit 1
+fi
+
 mv luga-dummy/* .
 rmdir luga-dummy
